made the integration in signup route

+++ commonStructs/UserKeyMethods.go
@@ -8,7 +8,7 @@ import (
 // we are not checking if the user is free tier, this method just copies it and calls it a day, and will error if the user in db is empty,
 //
 //hard code values/assumptions: the user tier will be free and the version to be 0, and gthe time to check for update on is after 1 month and 1 day
-func  (userKey *UserKey) InitializeTheStructForNewUser(  userInDb UserInDb , primaryKeyOfTheUserReturnedFromTheDB int)error{
+func  (userKey *UserKey) InitializeTheStructForNewUser(  userInDb UserInDb , primaryKeyOfTheUserReturnedFromTheDB int64)error{
 	userInDb.AddUserToFreeTier()
 	if !userInDb.IsUserValid(){
 		return fmt.Errorf("the user struct is not valid ")
@@ -19,7 +19,7 @@ func  (userKey *UserKey) InitializeTheStructForNewUser(  userInDb UserInDb , pri
 	userKey.IsUserPaid = userInDb.IsUserPaid
 	userKey.UserTier = userInDb.UserTeir
 	userKey.Version = 0
-	userKey.IDPrimaryKey = int64(primaryKeyOfTheUserReturnedFromTheDB)
+	userKey.IDPrimaryKey = primaryKeyOfTheUserReturnedFromTheDB
 	userKey.CheckForKeyUpdateOn = time.Now().AddDate(0,1,1).Unix()
 	return nil
 }
